imtiaz-mart/product-services/pyproject.toml
[project]
name = "product-services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "sqlmodel>=0.0.24",
]
